test(core): move the direct-light documentation coverage to repository level
The architecture-invariant and swept-claim checks read docs/architecture/ and
several guides/ and seeds/ paths, so they reached above packs/core via
parents[5]. The pack-tests-stay-in-pack boundary rejects that, correctly: a
pack's tests ship with the pack, so a test that depends on repository layout
passes here and fails in an adopter checkout where those paths do not exist.

Both cases move to tests/roster/, alongside the existing repository-level
work-loop and work-intake coverage. The pack-local test keeps only its
pack-anchored SKILL.md reads. The new file records why it lives at repo level so
the coverage is not moved back.

Caught by make build-check's boundary lint, not by the focused suites: the tests
passed where they were, they were simply owned by the wrong layer.
